backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
httpx==0.27.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pandas==2.2.1
jinja2==3.1.3
python-multipart==0.0.9
google-generativeai==0.4.1
